Moved Movement Methods to Movement.cs
